chore: Rename utilities
These are becoming pretty core parts of the project, lets give them a consistent NestJS-style naming scheme
