+++ server/src/@common/dto/run/runs.dto.ts
@@ -5,7 +5,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { IsDateString, IsDefined, IsInt, IsOptional, ValidateNested } from 'class-validator';
 import { Expose } from 'class-transformer';
 import { MapDto } from '../map/map.dto';
-import { DtoTransform } from '../../utils/dto-utils';
+import { DtoTransform } from '../../utils/dto.utility';
 
 // TODO: BaseStatsDTO, various other nested DTOs
