+++ server/src/@common/utils/bitflag.utility.ts
